+++ test/input/func_noerror_proteted_meta_access.py
@@ -19,4 +19,5 @@ class ModelWhichLikesMeta(models.Model):
 if __name__ == '__main__':
     MODEL = ModelWhichLikesMeta()
     MODEL.save()
+    print MODEL._meta.get_field('ursuary')
     print MODEL.do_a_thing()